Refactor RecommendHomeContainer component and update import path for PrimaryRecommondationCard
